Lua stuff add.
I create a new exe which has access to lua and link that with ldns.
I need to stub some ldns functions to see if it will actually work.
You can now "program" ldns from a script which will be run when you
execute lua-rns.

Examples is added which calucales an average (not a ldns functions
of course).