Sorry, previous commit (revision 3494, files: dnssec.c, dnssec_sign.c, examples/ldns-verify-zone.c, ldns/dnssec.h, ldns/dnssec_sign.h and ldns_symbols.def) should have included this documentation change and the following text:
Added the ldns_dnssec_mark_and_get_glue function to inventorize the glue from a ldns_dnssec_zone structure.
Also modified the marking to mark all names that are occluded and not just the names that contain only glue.
Modified the signing, adding of nsec and nsec3 functions to handle the ommission of occluded names and obscured rrsets (in stead of only handling the ommision of glue).
Modified the documentation to clarify how ldns_dnssec_mark_glue and ldns_dnssec_name_node_next_nonglue are intended to work togetger.

See also bug 380.