pyLDNS: Fixed mem leaks, changed behaviour of ldns_resolver.trusted_key().
Method ldns_resolver.trusted_key() now returns a rr list of trusted key copies.
This is because the underlying C function does not create a list of
clones but only creates adds pointers to existing list. This causes double free
errors when freeing the lists in Python. Blind cloning of list content
causes memory leaks because of the nature of the wrapped function.
Therefore the wrapper has to ensure that the list of trusted keys is
created from scratch every time the Python method is called.