patch for ldns-dane
Hi,

someone over here needed a little addition to ldns-dane; he wanted it to
be able to use a different nameserver than the system-configured one.

Attached patch adds a '-r <ip address>' option to ldns-dane, and uses
said ip address as the resolver instead of the resolvers specified in
resolv.conf.

Please consider this for inclusion in a next release :)

Cheers,

Jelte