Add support for building with CMake.
This change adds the ability to configure and build the project via
CMake by providing a set of checks that attempt to reproduce the
behaviour of the autotools scripts as closely as possible. Also removed
an errant semicolon in util.c

Additionally, a new gettimeofday.c source file has been added which
enables the code to function. The code has been copied from
http://doxygen.postgresql.org/gettimeofday_8c_source.html

It has been tested on Mac, Linux and Windows.

This change and all subsequent changes by myself are hereby released
under the terms contained in LICENSE

Signed-off-by: Oliver Ddin <oddin@mimecast.com>