Patches for the parsing bugs
Hi Willem,

The attached patch appears to suffice to resolve the three bugs I've
found so far.

I'm not saying it's a perfect fix, but with this in place I didn't
manage to trigger any more crashes yet.

I've started fuzzing on the wire format parser now, so far no news!

Ray